+++ Assets/Novel/Runtime/NovelScenarioRunner.cs
@@ -59,6 +59,7 @@ namespace Novel.Runtime
                 config.AddCommand<PortraitCommand>("portrait");
                 config.AddCommand<StageCommand>("stage");
                 config.AddCommand<ExitCommand>("exit");
+                config.AddCommand<SpeakerCommand>("speaker");
                 config.AddCommand<ClearStageCommand>("clear_stage");
                 config.AddCommand<BackgroundCommand>("bg");
                 config.AddCommand<StillCommand>("still");